+++ lib-mal/CHANGELOG.md
@@ -10,3 +10,5 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Oauth2 authorization 
 - `Get anime list` API funciton
 - `Get anime details` API function
+- `Get anime ranking` API function
+- Enums for available fields and request options